close sidebar after logout
